Added some vars for future use, camera fucntionality, and also a slight enhancement
I want to talk to you about how we want to do characters, and apply the code to the gameobjects. Besides adding a couple vars and lookign through different things, i added nothing.

The camera functionality comes from the playerController class, essentially the camera will be how each client interactes with the world.

Also press tab when you are in the game